smaller weight float size

+++ bayes_filter/misc.py
@@ -52,9 +52,9 @@ def maybe_create_posterior_solsets(datapack: DataPack, solset: str, posterior_na
                                 dir=patch_names)
             datapack.add_soltab('clock000', weightDtype='f64', time=times.mjd * 86400., pol=pol_labels,
                                 ant=antenna_labels)
-            datapack.add_soltab('phase000', weightDtype='f64', freq=freqs, time=times.mjd * 86400., pol=pol_labels,
+            datapack.add_soltab('phase000', weightDtype='f16', freq=freqs, time=times.mjd * 86400., pol=pol_labels,
                                 ant=antenna_labels, dir=patch_names)
-            datapack.add_soltab('amplitude000', weightDtype='f64', freq=freqs, time=times.mjd * 86400., pol=pol_labels,
+            datapack.add_soltab('amplitude000', weightDtype='f16', freq=freqs, time=times.mjd * 86400., pol=pol_labels,
                                 ant=antenna_labels, dir=patch_names)
 
             datapack.add_solset(screen_solset,
@@ -66,11 +66,10 @@ def maybe_create_posterior_solsets(datapack: DataPack, solset: str, posterior_na
                                 dir=screen_patch_names)
             datapack.add_soltab('clock000', weightDtype='f64', time=times.mjd * 86400., pol=pol_labels,
                                 ant=antenna_labels)
-            datapack.add_soltab('phase000', weightDtype='f64', freq=freqs, time=times.mjd * 86400., pol=pol_labels,
+            datapack.add_soltab('phase000', weightDtype='f16', freq=freqs, time=times.mjd * 86400., pol=pol_labels,
                                 ant=antenna_labels, dir=screen_patch_names)
-            datapack.add_soltab('amplitude000', weightDtype='f64', freq=freqs, time=times.mjd * 86400., pol=pol_labels,
+            datapack.add_soltab('amplitude000', weightDtype='f16', freq=freqs, time=times.mjd * 86400., pol=pol_labels,
                                 ant=antenna_labels, dir=screen_patch_names)
-
             datapack.current_solset = solset
 
 def dict2namedtuple(d, name="Result"):
